Transaction api

+++ backend/payment/src/main/java/com/dbs/payment/dao/EmployeeRepository.java
@@ -1,9 +1,11 @@
 package com.dbs.payment.dao;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.dbs.payment.model.Employee;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, BigInteger> {
 
 }
